Clip out of spec values for miter limit in BasicStroke.
The BasicStroke Constructor states that 1f is the mimimum value for the miter limit.
But BasicStroke does not validate and enforce this. It mostly seems to work.

For whatever reasons all BasicStroke attributes map 1:1 to the PDF spec. But in the
PDF spec this minimum value is not stated. But the Acrobat Reader will fail to
render the PDF if the miter limit < 1f.

The fix here is to do a Math.max(1f,miterLimit), so that the miter limit is always >= 1f.

This fixes #29.